add buildHistoryManager controller; deprecate buildHistory
https://coveord.atlassian.net/browse/KIT-466
