+++ packages/headless/src/controllers/history/headless-history.ts
@@ -1,49 +1,31 @@
 import {Engine} from '../../app/headless-engine';
-import {buildController} from '../controller/headless-controller';
-import {back, forward} from '../../features/history/history-actions';
-import {executeSearch} from '../../features/search/search-actions';
-import {
-  logNavigateBackward,
-  logNavigateForward,
-} from '../../features/history/history-analytics-actions';
+import {buildHistoryManager} from './headless-history-manager';
 
 /**
  * The `History` controller is in charge of allowing navigating back and forward in the search interface history.
+ *
+ * @deprecated The `History` type alias will be removed in a future release. Please use `HistoryManager` instead.
  */
 export type History = ReturnType<typeof buildHistory>;
-/** The state relevant to the `History` controller.*/
+
+/**
+ * The state relevant to the `History` controller.
+ *
+ * @deprecated The `HistoryState` type alias will be removed in a future release. Please use `HistoryManagerState` instead.
+ * */
 export type HistoryState = History['state'];
 
+/**
+ * Creates a `History` controller instance.
+ *
+ * @param engine - The headless engine.
+ * @returns A `History` controller instance.
+ *
+ * @deprecated The `buildHistory` controller will be removed in a future release. Please use `buildHistoryManager` instead.
+ */
 export const buildHistory = (engine: Engine) => {
-  const controller = buildController(engine);
-  const {dispatch} = engine;
-
-  return {
-    ...controller,
-    get state() {
-      return engine.state.history;
-    },
-
-    /**
-     * Move backward in the interface history.
-     */
-    async back() {
-      if (!this.state.past.length || !this.state.present) {
-        return;
-      }
-      await dispatch(back());
-      dispatch(executeSearch(logNavigateBackward()));
-    },
-
-    /**
-     * Move forward in the interface history.
-     */
-    async forward() {
-      if (!this.state.future.length || !this.state.present) {
-        return;
-      }
-      await dispatch(forward());
-      dispatch(executeSearch(logNavigateForward()));
-    },
-  };
+  engine.logger.error(
+    'The "buildHistory" controller will be removed in a future release. Please use "buildHistoryManager" instead.'
+  );
+  return buildHistoryManager(engine);
 };
